Rename test with type worksapce -> workspace
