+++ tests/e2e/tests/devfiles/JavaMaven.spec.ts
@@ -11,7 +11,7 @@ import 'reflect-metadata';
 import { WorkspaceNameHandler} from '../..';
 import * as projectAndFileTests from '../../testsLibrary/ProjectAndFileTests';
 import * as commonLsTests from '../../testsLibrary/LsTests';
-import * as workspaceHandling from '../../testsLibrary/WorksapceHandlingTests';
+import * as workspaceHandling from '../../testsLibrary/WorkspaceHandlingTests';
 import * as codeExecutionTests from '../../testsLibrary/CodeExecutionTests';
 
 const workspaceSampleName: string = 'console-java-simple';
