+++ tests/e2e/tests/devfiles/NodeJS.spec.ts
@@ -12,7 +12,7 @@ import 'reflect-metadata';
 import * as codeExecutionHelper from '../../testsLibrary/CodeExecutionTests';
 import * as commonLsTests from '../../testsLibrary/LsTests';
 import * as projectManager from '../../testsLibrary/ProjectAndFileTests';
-import * as workspaceHandling from '../../testsLibrary/WorksapceHandlingTests';
+import * as workspaceHandling from '../../testsLibrary/WorkspaceHandlingTests';
 
 const workspaceStack: string = 'NodeJS Express Web Application';
 const workspaceSampleName: string = 'nodejs-web-app';
